server/models/ecot/requirements.txt
timm==0.9.10
tokenizers==0.19.1
torch>=2.2.0
torchvision>=0.16.0
transformers==4.40.1
transforms3d
matplotlib
numpy
pillow
opencv-python-headless
fastapi
uvicorn
pydantic
accelerate
python-multipart